server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain==0.1.16
langchain-cli==0.0.21
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langcodes==3.4.0
langserve==0.1.0
langsmith==0.1.49
language_data==1.2.0
lxml==5.2.1
lxml_html_clean==0.1.1
marisa-trie==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.23.2
orjson==3.10.1
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.13
pydantic_core==2.18.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==26.0.2
regex==2024.4.16
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.6.0
tinysegmenter==0.3
tldextract==5.1.2
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.23.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
yarl==1.9.4
